+++ package.json
@@ -34,6 +34,7 @@
     "ol": "^10.6.1",
     "react": "^19.1.0",
     "react-chartjs-2": "^5.3.0",
-    "react-dom": "^19.1.0"
+    "react-dom": "^19.1.0",
+    "xlsx": "^0.18.5"
   }
 }
