+++ src/routes/application.tsx
@@ -1,11 +1,10 @@
 import { Link, Outlet } from "@tanstack/react-router";
 import React, { useMemo, useState } from "react";
 import { ApplicationContext } from "../applicationContext";
-import {
-  calculateTotalYearlyOilProduction,
-  generateCompleteData,
-} from "../utils/projections";
+import { generateCompleteData } from "../utils/projections";
+import { calculateTotalYearlyOilProduction } from "../utils/calculations";
 import { data } from "../generated/data";
+import { exportDataSet } from "../utils/excel";
 
 export function Application() {
   const [year, setYear] = useState(2025);
@@ -44,6 +43,9 @@ export function Application() {
         </div>
       </header>
       <main>
+        <button onClick={() => exportDataSet(fullData)}>
+          Last ned som Excel
+        </button>
         <Outlet />
       </main>
     </ApplicationContext>
